+++ packages/app/src/cli/prompts/dev.test.ts
@@ -9,11 +9,13 @@ import {
 } from './dev.js'
 import {Organization, OrganizationStore} from '../models/organization.js'
 import {testOrganizationApp} from '../models/app/app.test-data.js'
-import {describe, expect, vi, test} from 'vitest'
+import {getTomls} from '../utilities/app/config/getTomls.js'
+import {describe, expect, vi, test, beforeEach} from 'vitest'
 import {renderAutocompletePrompt, renderConfirmationPrompt, renderTextPrompt} from '@shopify/cli-kit/node/ui'
 import {mockAndCaptureOutput} from '@shopify/cli-kit/node/testing/output'
 
 vi.mock('@shopify/cli-kit/node/ui')
+vi.mock('../utilities/app/config/getTomls')
 
 const ORG1: Organization = {
   id: '1',
@@ -46,6 +48,10 @@ const STORE2: OrganizationStore = {
   transferDisabled: false,
   convertableToPartnerTest: false,
 }
+
+beforeEach(() => {
+  vi.mocked(getTomls).mockResolvedValue({})
+})
 
 describe('selectOrganization', () => {
   test('request org selection if passing more than 1 org', async () => {
@@ -95,6 +101,29 @@ describe('selectApp', () => {
       choices: [
         {label: 'app1', value: 'key1'},
         {label: 'app2', value: 'key2'},
+      ],
+      search: expect.any(Function),
+      hasMorePages: true,
+    })
+  })
+
+  test('includes toml names when present', async () => {
+    vi.mocked(getTomls).mockResolvedValueOnce({
+      [APP1.apiKey]: 'shopify.app.toml',
+      [APP2.apiKey]: 'shopify.app.dev.toml',
+    })
+
+    const apps = {nodes: [APP1, APP2], pageInfo: {hasNextPage: true}}
+    vi.mocked(renderAutocompletePrompt).mockResolvedValue('key2')
+
+    const got = await selectAppPrompt(apps, ORG1.id, 'token', {directory: '/'})
+
+    expect(got).toEqual(APP2.apiKey)
+    expect(renderAutocompletePrompt).toHaveBeenCalledWith({
+      message: 'Which existing app is this for?',
+      choices: [
+        {label: 'app1 (shopify.app.toml)', value: 'key1'},
+        {label: 'app2 (shopify.app.dev.toml)', value: 'key2'},
       ],
       search: expect.any(Function),
       hasMorePages: true,
